[docs] Add Realm ABI Specification
The Realm ABI Specification formally describes the protocol
and encoding of interactions between realms and external clients,
independently of their language.

It can be seen as an extension of the
[ABCI](https://github.com/tendermint/tendermint/tree/master/spec/abci)
which covers the protocol between blockchain nodes and clients, but
without details on blockchain practical applications, such as realms.

This PR addresses [#1842].

This is an initial draft commit, work in progress.